Renames + refactoring some parts
